create clean migration with server default autoincrementing publi id on document
