+++ backend/app/alembic/versions/c859c1ba61a0_document_public_id.py
@@ -1,0 +1,55 @@
+"""document public id
+
+Revision ID: c859c1ba61a0
+Revises: 239ca64ca70d
+Create Date: 2025-07-25 14:32:12.728476
+
+"""
+
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = "c859c1ba61a0"
+down_revision: Union[str, None] = "239ca64ca70d"
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # Create sequence first
+    op.execute("CREATE SEQUENCE IF NOT EXISTS document.document_public_id_seq")
+    # Adjusted to add server default
+    op.add_column(
+        "document",
+        sa.Column(
+            "public_id",
+            sa.Integer(),
+            autoincrement=True,
+            nullable=False,
+            server_default=sa.text("nextval('document.document_public_id_seq')"),
+        ),
+        schema="document",
+    )
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_index(
+        op.f("ix_document_document_public_id"),
+        "document",
+        ["public_id"],
+        unique=True,
+        schema="document",
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(
+        op.f("ix_document_document_public_id"), table_name="document", schema="document"
+    )
+    op.drop_column("document", "public_id", schema="document")
+    op.execute("DROP SEQUENCE IF EXISTS document.document_public_id_seq")
+    # ### end Alembic commands ###
